+++ include/tools/regexTools.h
