+++ include/tools/stringTools.h
