+++ src/book.cpp
@@ -20,9 +20,9 @@
 #include "book.h"
 #include "reader.h"
 
-#include "common/base64.h"
-#include "common/regexTools.h"
-#include "common/networkTools.h"
+#include "tools/base64.h"
+#include "tools/regexTools.h"
+#include "tools/networkTools.h"
 
 #include <pugixml.hpp>
