+++ src/opds_dumper.cpp
@@ -20,7 +20,7 @@
 #include "opds_dumper.h"
 #include "book.h"
 
-#include <common/otherTools.h>
+#include <tools/otherTools.h>
 
 namespace kiwix
 {
