+++ src/tools/otherTools.cpp
@@ -17,7 +17,7 @@
  * MA 02110-1301, USA.
  */
 
-#include <common/otherTools.h>
+#include <tools/otherTools.h>
 #include <map>
 
 static std::map<std::string, std::string> codeisomapping {
